Generalize hash_to_field
Move hash_to_field outside of sha3 mod.
Set proptest back to unit test.
TODO: remove sha3 dependency.
