+++ crypto/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2021"
 path = "../math"
 
 [dependencies]
-proptest = "1.1.0"
 sha3 = "0.10.6"
 
 [dependencies.rand]
