+++ crypto/src/hash/helpers.rs
