+++ crypto/src/hash/mod.rs
@@ -1,3 +1,5 @@
+pub mod hash_to_field;
+pub mod helpers;
 pub mod poseidon;
 pub mod sha3;
 pub mod traits;
